question 01 finalized (no doxy yet)

+++ src/q01/volume.cpp
@@ -2,17 +2,17 @@
 #include "volume.h"
 #define PI 3.14
 
-float vol_pyramid(float base_area, float heigth){
+float vol_pyramid(int base_area, int heigth){
 	return 1/3*(base_area*heigth);
 }
 
-float vol_cube(float edge){
+float vol_cube(int edge){
 	return (edge*edge*edge);
 }
 
-float vol_parallelepiped(float edge1, float edge2, float edge3){
+float vol_parallelepiped(int edge1, int edge2, int edge3){
 	return edge1*edge2*edge3;
 }
-float vol_sphere(float r){
+float vol_sphere(int r){
 	return 4/3*PI*(r*r*r);
 }
